Consider some ideas around how the code would actually look if both primitive streams and object streams were parsed from the same source.
Feature parity and name correspondence would be the defining feature of this.
I had an idea around working the transpiled and native versions toward a unified middle ground.

This, in abstract, is a nice idea.
If the versions do not match, either one is wrong (or the bespoke version is off inreplicable complexity, which is perhaps a bad thing).

If the transpiled version is wrong, fix the transpiler to generate the correct result.
If the native source is wrong, use the code from the transpiled version instead.

In reality, the comparison function is going to need to be some sort of AST diff based thing.
For the target language.

!!Scratch that, I'm an idiot.

Given that one of the fundamental goals that I have here is maximally simplistic compiler production, I will only be writing parsers for languages that are of a significantly less ambiguous token form than my current target language of java.

Just run through the damned streams of generated and native code byte-wise ensuring equality.
It doesn't matter that you don't understand the code on more than a line by line basis.

Anyway, there's a lot of other shit in here too, debugs, rewrites, api refactorings.

It's about time to actually commit.