Do a bad refactoring.  - Introduce contexts universally in the parsers.   - This isn't used for anything yet, and so the default was simply to use a NullContext.    - But even this isn't done overly well.     - It's inconsistent, and the API would be smarter to accommodate for the notion of a ContextFreeHandler.     - This would override the default method, and call through to a method call lacking the context.     - The code of the ContextFreeHandler implementations would be significantly closer to the code before this change, and so much of this will be rolled back.     - I'll enforce the (compile time) correctness of the revert by genericising the Handler type.
However, this was an absolutely brain-free code change - the reverse will be a little more thought out.
It was done simply to allow for touched tail detection and for a context to be passed into the lineHandlers parsers.
The touched tail detection makes for parsers such as the not yet implemented new version of the tree parser somewhat cleaner to implement.
The context that needs to be passed into the parser is extracted from the indent count of the line-head.
It can then be fed down the context aware processing chain, with the handler that empties the head also being the direct parent of the children in the indent tree.

This also needs to be revalidated by reimplementing the reconstitute functionality.